requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.10.1
atpublic==3.0.1
attrs==21.4.0
backports.lzma==0.0.14
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cmake==3.22.4
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==36.0.2
cvxopt==1.3.0
cvxpy==1.2.0
dagshub==0.1.8
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
dpath==2.0.6
dulwich==0.20.35
dvc==2.10.1
dvc-render==0.0.4
ecos==2.0.10
flatten-dict==0.4.2
flufl.lock==7.0
frozenlist==1.3.0
fsspec==2022.3.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
grandalf==0.6
idna==3.3
joblib==1.1.0
mailchecker==4.1.16
multidict==6.0.2
nanotime==0.5.2
networkx==2.8
numpy==1.22.1
osqp==0.6.1
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
phonenumbers==8.12.47
Pillow==9.0.1
psutil==5.9.0
pybind11==2.9.2
pycparser==2.21
pydot==1.4.2
pygit2==1.9.1
Pygments==2.12.0
pygtrie==2.4.2
pyparsing==3.0.8
python-benedict==0.25.0
python-dateutil==2.8.2
python-fsutil==0.6.0
python-slugify==6.1.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rich==12.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.7.3
scmrepo==0.0.16
scs==3.2.0
shortuuid==1.0.8
shtab==1.5.4
six==1.16.0
sklearn==0.0
smmap==5.0.0
tabulate==0.8.9
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
voluptuous==0.13.1
wcwidth==0.2.5
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
